first pass at Record.write
