+++ packages/houdini/src/runtime/public/record.ts
@@ -1,10 +1,17 @@
 import { rootID } from '../cache/cache'
 import type { TypeInfo } from '../cache/schema'
 import { keyFieldsForType } from '../lib/config'
-import type { SubscriptionSelection } from '../lib/types'
+import type { FragmentArtifact, GraphQLObject, SubscriptionSelection } from '../lib/types'
 import type { Cache } from './cache'
 import { _typeInfo } from './cache'
-import type { ArgType, CacheTypeDef, FieldType, TypeFieldNames, ValidTypes } from './types'
+import type {
+	ArgType,
+	CacheTypeDef,
+	FieldType,
+	FragmentList,
+	TypeFieldNames,
+	ValidTypes,
+} from './types'
 
 export class Record<Def extends CacheTypeDef, Type extends ValidTypes<Def>> {
 	#id: string
@@ -222,6 +229,18 @@ export class Record<Def extends CacheTypeDef, Type extends ValidTypes<Def>> {
 		return (Array.isArray(data) ? finalResult : finalResult[0]) as FieldType<Def, Type, Field>
 	}
 
+	write<_Fragment extends { artifact: FragmentArtifact }>(args: {
+		fragment: _Fragment
+		data: ListKeyValue<FragmentList<Def, Type>, _Fragment>
+	}) {
+		// we have the data and the fragment, just pass them both to the cache
+		this.#cache._internal_unstable.write({
+			data: args.data as unknown as GraphQLObject,
+			selection: args.fragment.artifact.selection,
+			parent: this.#id,
+		})
+	}
+
 	delete() {
 		this.#cache._internal_unstable.delete(this.#id)
 	}
@@ -325,3 +344,11 @@ function unmarshalNestedList<Def extends CacheTypeDef>(cache: Cache<Def>, list: 
 
 	return newValue
 }
+
+type ListKeyValue<List, _Target> = List extends [infer Head, ...infer Rest]
+	? Head extends [infer _Key, infer _Value]
+		? _Key extends _Target
+			? _Value
+			: ListKeyValue<Rest, _Target>
+		: 'Encountered unknown fragment'
+	: 'Encountered unknown fragment'
