+++ packages/houdini/src/runtime/public/tests/record.test.ts
@@ -1,7 +1,7 @@
 import { test, expect } from 'vitest'
 
 import { rootID } from '../../cache/cache'
-import type { SubscriptionSelection } from '../../lib'
+import { ArtifactKind, FragmentArtifact, SubscriptionSelection } from '../../lib'
 import { marshalNestedList, Record } from '../record'
 import { testCache } from './test'
 
@@ -481,4 +481,99 @@ test('can set nested lists of record proxies', function () {
 
 	// and make sure we get the right value from the imperative API
 	expect(marshalNestedList(cache.root.get({ field: 'listOfLists' }))).toEqual(expected)
+})
+
+test('can writeFragments', function () {
+	const cache = testCache()
+
+	const selection: SubscriptionSelection = {
+		fields: {
+			viewer: {
+				type: 'User',
+				keyRaw: 'viewer',
+				selection: {
+					fields: {
+						id: {
+							type: 'ID',
+							keyRaw: 'id',
+						},
+						firstName: {
+							type: 'String',
+							keyRaw: 'firstName',
+						},
+						__typename: {
+							type: 'String',
+							keyRaw: '__typename',
+						},
+						parent: {
+							type: 'User',
+							keyRaw: 'parent',
+							selection: {
+								fields: {
+									id: {
+										type: 'ID',
+										keyRaw: 'id',
+									},
+									firstName: {
+										type: 'String',
+										keyRaw: 'firstName',
+									},
+									__typename: {
+										type: 'String',
+										keyRaw: '__typename',
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	// write the data as a deeply nested object
+	cache._internal_unstable.write({
+		selection,
+		data: {
+			viewer: {
+				id: '1',
+				firstName: 'bob',
+				__typename: 'User',
+				parent: {
+					id: '2',
+					firstName: 'jane',
+					__typename: 'User',
+				},
+			},
+		},
+	})
+
+	const artifact: FragmentArtifact = {
+		kind: ArtifactKind.Fragment,
+		name: 'string',
+		raw: 'string',
+		hash: 'string',
+		rootType: 'string',
+		selection: {
+			fields: {
+				firstName: {
+					type: 'String',
+					keyRaw: 'firstName',
+				},
+			},
+		},
+	}
+
+	// write a fragment to update User:2
+	cache.get('User', { id: '2' }).write({
+		fragment: {
+			artifact,
+		},
+		data: {
+			firstName: 'michael',
+		},
+	})
+
+	// make sure we updated the field
+	expect(cache.get('User', { id: '2' }).get({ field: 'firstName' })).toEqual('michael')
 })
