+++ packages/houdini/src/runtime/public/tests/test.ts
@@ -1,5 +1,6 @@
 import { testConfigFile } from '../../../test'
 import { Cache as _Cache } from '../../cache/cache'
+import { FragmentArtifact } from '../../lib'
 import { Cache } from '../cache'
 import type { Record } from '../record'
 
@@ -8,6 +9,7 @@ type CacheTypeDef = {
 	types: {
 		__ROOT__: {
 			idFields: {}
+			fragments: []
 			fields: {
 				test: {
 					type: number | null
@@ -53,6 +55,7 @@ type CacheTypeDef = {
 			idFields: {
 				id: string
 			}
+			fragments: [[{ artifact: FragmentArtifact }, { firstName: string }]]
 			fields: {
 				firstName: {
 					type: string
@@ -76,6 +79,7 @@ type CacheTypeDef = {
 			idFields: {
 				id: string
 			}
+			fragments: []
 			fields: {
 				name: {
 					type: string | null
@@ -99,6 +103,7 @@ type CacheTypeDef = {
 			idFields: {
 				id: string
 			}
+			fragments: []
 			fields: {
 				name: {
 					type: string | null
